+++ src/main/java/gwasuwonshot/tutice/lesson/controller/LessonController.java
@@ -3,8 +3,8 @@ package gwasuwonshot.tutice.lesson.controller;
 import gwasuwonshot.tutice.common.exception.SuccessStatus;
 import gwasuwonshot.tutice.common.resolver.userIdx.UserIdx;
 import gwasuwonshot.tutice.common.dto.ApiResponseDto;
-import gwasuwonshot.tutice.lesson.dto.request.CreateLessonMaintenanceRequestDto;
-import gwasuwonshot.tutice.lesson.dto.request.CreateLessonRequestDto;
+import gwasuwonshot.tutice.lesson.dto.request.createLesson.CreateLessonMaintenanceRequestDto;
+import gwasuwonshot.tutice.lesson.dto.request.createLesson.CreateLessonRequestDto;
 import gwasuwonshot.tutice.lesson.dto.request.UpdateLessonParentsRequestDto;
 import gwasuwonshot.tutice.lesson.dto.response.CreateLessonResponseDto;
 import gwasuwonshot.tutice.lesson.dto.response.GetLessonByUserResponseDto;
