+++ src/main/java/gwasuwonshot/tutice/lesson/dto/request/createLesson/CreateLessonRequestDto.java
@@ -1,4 +1,4 @@
-package gwasuwonshot.tutice.lesson.dto.request;
+package gwasuwonshot.tutice.lesson.dto.request.createLesson;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AccessLevel;
@@ -6,10 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.validation.Valid;
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Pattern;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
